Struct oxcable_basic_devices::delay::Delay

pub struct Delay {
    // some fields omitted
}

A simple delay filter.

Methods

impl Delay

fn new(delay: f32, feedback: f32, wetness: f32, num_channels: usize) -> Self

Creates a new Delay object.

delay: the time to delay the input, in seconds

feedback: how much of our delayed signal to feed back into the next delay; should be between 0.0 and 1.0.

wetness: how much of our input signal to mix into the delayed signal in the output; should be between 0.0 and 1.0.