Module oxcable_basic_devices::dynamics

Audio devices to perform dynamics processing.

Dynamics processors use an estimate of the signal envelope to modify the dynamics of the final signal. The LevelDetector can be used to track the peak amplitude of a signal over time, and is uesd by the other three filters.

Reexports

pub use self::compressor::Compressor;

pub use self::limiter::Limiter;

pub use self::noise_gate::NoiseGate;

Modules

compressor
A compression filter.

limiter
A limiter.

noise_gate
A noise gate.