Struct oxcable_basic_devices::dynamics::noise_gate::NoiseGate

pub struct NoiseGate {
    // some fields omitted
}

A noise gate.

Noise gates provide a floor the signal must exceed; anything below the floor is muted instead.

The noise gate provides two different thresholds. The gate will open once the signal exceeds the on threshold, and close when the signal level drops below the off threshold.

Methods

impl NoiseGate

fn new(on_threshold: f32, off_threshold: f32, gain: f32, num_channels: usize) -> Self

Creates a new compressor.

The on_threshold and off_threshold specify the signal floors in decibels.

The specified gain (in decibels) will be applied to the signal after compression.